+++ Private/Save-HelmBackup.ps1
@@ -6,7 +6,7 @@ function Save-HelmBackup {
         [switch]$Verbose,
         [switch]$AllNamespaces,
         [switch]$AllNonSystemNamespaces,
-        [switch]$SnapshotHelm,           # Perform a full Helm snapshot
+        [switch]$SnapshotHelm, # Perform a full Helm snapshot (includes backend storage)
         [switch]$SnapshotHelmUsedValues # Perform only Helm used values snapshot
     )
 
@@ -75,10 +75,46 @@ function Save-HelmBackup {
 
         if ($usedValuesOutput) {
             return $usedValuesOutput
-        } else {
+        }
+        else {
             Write-Host "No used values found for release '$ReleaseName'." -ForegroundColor Yellow
             return $null
         }
+    }
+
+    # Function to backup backend storage
+    function Backup-HelmBackend {
+        param (
+            [string]$Namespace
+        )
+        Write-Verbose "Backing up Helm storage backend for namespace: $Namespace"
+        $secrets = kubectl get secrets -n $Namespace -o json | ConvertFrom-Json
+        $backendType = $secrets.items | Where-Object { $_.type -eq 'helm.sh/release.v1' -and $_.metadata.name -like '*sh.helm.release.v1*' } | Select-Object -First 1
+
+        if ($backendType -match "helm.sh/release.v1") {
+            Write-Verbose "Detected backend: Secrets"
+            $releases = kubectl get secrets -n $Namespace -l "owner=helm" -o json | ConvertFrom-Json
+        }
+        else {
+            Write-Verbose "Detected backend: ConfigMaps"
+            $releases = kubectl get configmaps -n $Namespace -l "owner=helm" -o json | ConvertFrom-Json
+        }
+
+        # Prepare YAML output file
+
+        $releaseName = $release.metadata.labels.name
+        $safeName = $releaseName -replace "[:\\/]", "_"
+        $timestamp = Get-Date -Format "yyyy-MM-dd_HH-mm-ss"
+        $filePath = Join-Path -Path $OutputPath -ChildPath "${safeName}_backend_$timestamp.yaml"
+        New-Item -Path $filePath -ItemType File -Force | Out-Null
+
+        # Convert and append each release to the YAML file
+        foreach ($release in $releases.items) {
+            $yamlContent = ConvertTo-Yaml -Data $release
+            Add-Content -Path $filePath -Value $yamlContent
+            Add-Content -Path $filePath -Value "---"
+        }
+        Write-Host "Helm release '$releaseName' backend saved: $filePath" -ForegroundColor Green
     }
 
     # Function to process a namespace
@@ -90,6 +126,11 @@ function Save-HelmBackup {
         $namespaceOption = ""
         if ($Namespace) {
             $namespaceOption = "-n $Namespace"
+        }
+
+        # Back up backend storage if SnapshotHelm is enabled
+        if ($SnapshotHelm) {
+            Backup-HelmBackend -Namespace $Namespace
         }
 
         $helmListCmd = "list $namespaceOption --output json"
@@ -111,8 +152,9 @@ function Save-HelmBackup {
 
                 if ($DryRun) {
                     Write-Host "Dry run: Found Helm release '$releaseName' in namespace '$releaseNamespace'."
-                } else {
-                    # Full Helm snapshot (values, manifest, and history)
+                }
+                else {
+                    # Full Helm snapshot (values, and manifest)
                     if ($SnapshotHelm) {
                         # Fetch values
                         $helmGetValuesCmd = "get values $releaseName $namespaceOption -o yaml"
@@ -134,16 +176,6 @@ function Save-HelmBackup {
                             $manifestOutput | Out-File -FilePath $manifestFile -Force
                             Write-Host "Helm release '$releaseName' manifest saved: $manifestFile" -ForegroundColor Green
                         }
-
-                        # Fetch history
-                        $helmHistoryCmd = "history $releaseName $namespaceOption --output yaml"
-                        Write-Verbose "Running command: helm $helmHistoryCmd"
-                        $historyOutput = Invoke-HelmCommand $helmHistoryCmd
-                        if ($historyOutput) {
-                            $historyFile = Join-Path -Path $OutputPath -ChildPath "${safeReleaseName}_history_$timestamp.yaml"
-                            $historyOutput | Out-File -FilePath $historyFile -Force
-                            Write-Host "Helm release '$releaseName' history saved: $historyFile" -ForegroundColor Green
-                        }
                     }
 
                     # Only Helm used values snapshot
@@ -158,7 +190,8 @@ function Save-HelmBackup {
                     }
                 }
             }
-        } else {
+        }
+        else {
             Write-Host "No Helm releases found in the namespace '$Namespace'." -ForegroundColor Yellow
         }
     }
@@ -168,12 +201,15 @@ function Save-HelmBackup {
         if ($AllNamespaces) {
             $namespaces = kubectl get namespaces -o json | ConvertFrom-Json
             $allClusterNamespaces = $namespaces.items | ForEach-Object { $_.metadata.name }
-        } elseif ($AllNonSystemNamespaces) {
+        }
+        elseif ($AllNonSystemNamespaces) {
             $namespaces = kubectl get namespaces -o json | ConvertFrom-Json
             $allClusterNamespaces = $namespaces.items | Where-Object { $_.metadata.name -notmatch "^(kube-system|kube-public|kube-node-lease|default)$" } | ForEach-Object { $_.metadata.name }
-        } elseif ($Namespace) {
+        }
+        elseif ($Namespace) {
             $allClusterNamespaces = @($Namespace)
-        } else {
+        }
+        else {
             Write-Host "No namespace specified." -ForegroundColor Red
             return
         }